+++ priv/repo/migrations/20201105175242_create_transactions.exs
@@ -4,7 +4,7 @@ defmodule Stone.Repo.Migrations.CreateTransactions do
   def change do
     create table(:transactions, primary_key: false) do
       add :id, :binary_id, primary_key: true
-      add :amount, :float
+      add :amount, :integer
       add :type, :string
       add :destination, :string
       add :checking_account_id, references(:checking_accounts, on_delete: :nothing, type: :binary_id)
